requirements.txt
PyQt6==6.8.1
python-dateutil==2.9.0.post0
